requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0  # PostgreSQL async driver

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Configuration and environment
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Payment processing
stripe>=7.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
faker>=20.0.0

# Utility dependencies
python-decouple>=3.8
typing-extensions>=4.8.0

# Document Generation dependencies
ollama>=0.4.0
reportlab>=4.0.0
Pillow>=10.0.0
jinja2>=3.1.0